refactor(foreman): extract foreman AI into dedicated foreman/ package
Move all foreman-related code out of the monolithic main.py into a
structured package:

- backend/foreman/prompt.py   — FOREMAN_SYSTEM constant + build_system_prompt()
- backend/foreman/tools.py    — FOREMAN_TOOLS list, GitHub API helpers, exec_tools()
- backend/foreman/runner.py   — run_foreman_ai(), repair helpers, history serializer
- backend/foreman/state.py    — shared foreman_conversations dict (avoids circular imports)
- backend/foreman/__init__.py — public re-exports

WebSocket broadcast utilities (broadcast, emit_terminal_line, connections,
agent_owners) move to backend/events.py so both main.py and the foreman
package can import them without a circular dependency.

main.py shrinks from 2611 → 1540 lines (−41%). All 21 tests pass.

Closes #72
