+++ backend/foreman/__init__.py
@@ -1,0 +1,6 @@
+"""Foreman AI package — re-exports for convenient importing from main.py."""
+
+from foreman.runner import run_foreman_ai, serialize_history_for_debug
+from foreman.state import foreman_conversations
+
+__all__ = ["run_foreman_ai", "serialize_history_for_debug", "foreman_conversations"]
